requirements.txt
starlette_exporter
sqlalchemy
pydantic
psycopg2
python-slugify
pandas
openai
spacy